requirements.txt
Flask
Flask-Cors
SpeechRecognition
pydub
gtts
numpy
praat-parselmouth
gunicorn
